+++ sdk/compute/azure-mgmt-compute/generated_samples/gallery_image_version_delete_bypass_soft_delete.py
@@ -15,7 +15,7 @@ from azure.mgmt.compute import ComputeManagementClient
     pip install azure-identity
     pip install azure-mgmt-compute
 # USAGE
-    python gallery_image_update.py
+    python gallery_image_version_delete_bypass_soft_delete.py
 
     Before run the sample, please set the values of the client ID, tenant ID and client secret
     of the AAD application as environment variables: AZURE_CLIENT_ID, AZURE_TENANT_ID,
@@ -30,22 +30,14 @@ def main():
         subscription_id="SUBSCRIPTION_ID",
     )
 
-    response = client.gallery_images.begin_update(
+    client.gallery_image_versions.begin_delete(
         resource_group_name="myResourceGroup",
         gallery_name="myGalleryName",
         gallery_image_name="myGalleryImageName",
-        gallery_image={
-            "properties": {
-                "hyperVGeneration": "V1",
-                "identifier": {"offer": "myOfferName", "publisher": "myPublisherName", "sku": "mySkuName"},
-                "osState": "Generalized",
-                "osType": "Windows",
-            }
-        },
+        gallery_image_version_name="1.0.0",
     ).result()
-    print(response)
 
 
-# x-ms-original-file: 2026-03-03/galleryExamples/GalleryImage_Update.json
+# x-ms-original-file: 2026-03-03/galleryExamples/GalleryImageVersion_Delete_BypassSoftDelete.json
 if __name__ == "__main__":
     main()